"improvements from code review, cleanup non-implemented routes"
This reverts commit b6c422b334259b473949bce8e577e7e199530390.
