+++ libs/model-ad/api-client-angular/src/lib/.openapi-generator/FILES
@@ -1,13 +1,13 @@
 .gitignore
 README.md
 api.module.ts
-api/api-next-public-openapi-disease-correlation.service.ts
-api/api-next-public-openapi-gene-expression.service.ts
-api/api-next-public-openapi-model-overview.service.ts
-api/api-public-openapi-comparison-tool-config.service.ts
-api/api-public-openapi-data-version.service.ts
-api/api-public-openapi-model.service.ts
 api/api.ts
+api/comparison-tool-config.service.ts
+api/data-version.service.ts
+api/disease-correlation.service.ts
+api/gene-expression.service.ts
+api/model-overview.service.ts
+api/model.service.ts
 configuration.ts
 encoder.ts
 git_push.sh
