+++ libs/model-ad/api-client-angular/src/lib/api/data-version.service.ts
@@ -25,9 +25,7 @@ import { Observable } from 'rxjs';
 // @ts-ignore
 import { BasicError } from '../model/basic-error';
 // @ts-ignore
-import { ComparisonToolConfig } from '../model/comparison-tool-config';
-// @ts-ignore
-import { ComparisonToolPage } from '../model/comparison-tool-page';
+import { DataVersion } from '../model/data-version';
 
 // @ts-ignore
 import { BASE_PATH, COLLECTION_FORMATS } from '../variables';
@@ -36,7 +34,7 @@ import { Configuration } from '../configuration';
 @Injectable({
   providedIn: 'root',
 })
-export class ApiPublicOpenapiComparisonToolConfigService {
+export class DataVersionService {
   protected basePath = 'http://localhost/v1';
   public defaultHeaders = new HttpHeaders();
   public configuration = new Configuration();
@@ -109,14 +107,12 @@ export class ApiPublicOpenapiComparisonToolConfigService {
   }
 
   /**
-   * Get Comparison Tool configuration
-   * Retrieve the Comparison Tool configuration schema for the Model-AD application
-   * @param page Name of the page to retrieve the Comparison Tool configuration for
+   * Get data version
+   * Get data version
    * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
    * @param reportProgress flag to report request and response progress.
    */
-  public getComparisonToolConfig(
-    page: ComparisonToolPage,
+  public getDataVersion(
     observe?: 'body',
     reportProgress?: boolean,
     options?: {
@@ -124,9 +120,8 @@ export class ApiPublicOpenapiComparisonToolConfigService {
       context?: HttpContext;
       transferCache?: boolean;
     },
-  ): Observable<Array<ComparisonToolConfig>>;
-  public getComparisonToolConfig(
-    page: ComparisonToolPage,
+  ): Observable<DataVersion>;
+  public getDataVersion(
     observe?: 'response',
     reportProgress?: boolean,
     options?: {
@@ -134,9 +129,8 @@ export class ApiPublicOpenapiComparisonToolConfigService {
       context?: HttpContext;
       transferCache?: boolean;
     },
-  ): Observable<HttpResponse<Array<ComparisonToolConfig>>>;
-  public getComparisonToolConfig(
-    page: ComparisonToolPage,
+  ): Observable<HttpResponse<DataVersion>>;
+  public getDataVersion(
     observe?: 'events',
     reportProgress?: boolean,
     options?: {
@@ -144,9 +138,8 @@ export class ApiPublicOpenapiComparisonToolConfigService {
       context?: HttpContext;
       transferCache?: boolean;
     },
-  ): Observable<HttpEvent<Array<ComparisonToolConfig>>>;
-  public getComparisonToolConfig(
-    page: ComparisonToolPage,
+  ): Observable<HttpEvent<DataVersion>>;
+  public getDataVersion(
     observe: any = 'body',
     reportProgress: boolean = false,
     options?: {
@@ -155,17 +148,6 @@ export class ApiPublicOpenapiComparisonToolConfigService {
       transferCache?: boolean;
     },
   ): Observable<any> {
-    if (page === null || page === undefined) {
-      throw new Error(
-        'Required parameter page was null or undefined when calling getComparisonToolConfig.',
-      );
-    }
-
-    let localVarQueryParameters = new HttpParams({ encoder: this.encoder });
-    if (page !== undefined && page !== null) {
-      localVarQueryParameters = this.addToHttpParams(localVarQueryParameters, <any>page, 'page');
-    }
-
     let localVarHeaders = this.defaultHeaders;
 
     let localVarHttpHeaderAcceptSelected: string | undefined = options && options.httpHeaderAccept;
@@ -199,13 +181,12 @@ export class ApiPublicOpenapiComparisonToolConfigService {
       }
     }
 
-    let localVarPath = `/comparison-tool-config`;
-    return this.httpClient.request<Array<ComparisonToolConfig>>(
+    let localVarPath = `/data-version`;
+    return this.httpClient.request<DataVersion>(
       'get',
       `${this.configuration.basePath}${localVarPath}`,
       {
         context: localVarHttpContext,
-        params: localVarQueryParameters,
         responseType: <any>responseType_,
         withCredentials: this.configuration.withCredentials,
         headers: localVarHeaders,
