+++ libs/model-ad/api-client-angular/src/lib/api/gene-expression.service.ts
@@ -25,7 +25,9 @@ import { Observable } from 'rxjs';
 // @ts-ignore
 import { BasicError } from '../model/basic-error';
 // @ts-ignore
-import { DataVersion } from '../model/data-version';
+import { GeneExpression } from '../model/gene-expression';
+// @ts-ignore
+import { ItemFilterTypeQuery } from '../model/item-filter-type-query';
 
 // @ts-ignore
 import { BASE_PATH, COLLECTION_FORMATS } from '../variables';
@@ -34,7 +36,7 @@ import { Configuration } from '../configuration';
 @Injectable({
   providedIn: 'root',
 })
-export class ApiPublicOpenapiDataVersionService {
+export class GeneExpressionService {
   protected basePath = 'http://localhost/v1';
   public defaultHeaders = new HttpHeaders();
   public configuration = new Configuration();
@@ -107,12 +109,18 @@ export class ApiPublicOpenapiDataVersionService {
   }
 
   /**
-   * Get data version
-   * Get data version
+   * Get gene expression comparison data
+   * Returns a list of gene expression objects for use in comparison tools.
+   * @param category An ordered list of categories used to filter the data, where the first value is the category and the second is the subcategory. Pass each value by repeating the \&#39;category\&#39; query parameter, e.g. ?category&#x3D;category1&amp;category&#x3D;subcategoryA.
+   * @param item A list of items to filter the data by.
+   * @param itemFilterType The type of filter to apply to the items. Possible values are \&#39;include\&#39; or \&#39;exclude\&#39;.
    * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
    * @param reportProgress flag to report request and response progress.
    */
-  public getDataVersion(
+  public getGeneExpressions(
+    category: Array<string>,
+    item?: Array<string>,
+    itemFilterType?: ItemFilterTypeQuery,
     observe?: 'body',
     reportProgress?: boolean,
     options?: {
@@ -120,8 +128,11 @@ export class ApiPublicOpenapiDataVersionService {
       context?: HttpContext;
       transferCache?: boolean;
     },
-  ): Observable<DataVersion>;
-  public getDataVersion(
+  ): Observable<Array<GeneExpression>>;
+  public getGeneExpressions(
+    category: Array<string>,
+    item?: Array<string>,
+    itemFilterType?: ItemFilterTypeQuery,
     observe?: 'response',
     reportProgress?: boolean,
     options?: {
@@ -129,8 +140,11 @@ export class ApiPublicOpenapiDataVersionService {
       context?: HttpContext;
       transferCache?: boolean;
     },
-  ): Observable<HttpResponse<DataVersion>>;
-  public getDataVersion(
+  ): Observable<HttpResponse<Array<GeneExpression>>>;
+  public getGeneExpressions(
+    category: Array<string>,
+    item?: Array<string>,
+    itemFilterType?: ItemFilterTypeQuery,
     observe?: 'events',
     reportProgress?: boolean,
     options?: {
@@ -138,8 +152,11 @@ export class ApiPublicOpenapiDataVersionService {
       context?: HttpContext;
       transferCache?: boolean;
     },
-  ): Observable<HttpEvent<DataVersion>>;
-  public getDataVersion(
+  ): Observable<HttpEvent<Array<GeneExpression>>>;
+  public getGeneExpressions(
+    category: Array<string>,
+    item?: Array<string>,
+    itemFilterType?: ItemFilterTypeQuery,
     observe: any = 'body',
     reportProgress: boolean = false,
     options?: {
@@ -148,6 +165,39 @@ export class ApiPublicOpenapiDataVersionService {
       transferCache?: boolean;
     },
   ): Observable<any> {
+    if (category === null || category === undefined) {
+      throw new Error(
+        'Required parameter category was null or undefined when calling getGeneExpressions.',
+      );
+    }
+
+    let localVarQueryParameters = new HttpParams({ encoder: this.encoder });
+    if (category) {
+      category.forEach((element) => {
+        localVarQueryParameters = this.addToHttpParams(
+          localVarQueryParameters,
+          <any>element,
+          'category',
+        );
+      });
+    }
+    if (item) {
+      item.forEach((element) => {
+        localVarQueryParameters = this.addToHttpParams(
+          localVarQueryParameters,
+          <any>element,
+          'item',
+        );
+      });
+    }
+    if (itemFilterType !== undefined && itemFilterType !== null) {
+      localVarQueryParameters = this.addToHttpParams(
+        localVarQueryParameters,
+        <any>itemFilterType,
+        'itemFilterType',
+      );
+    }
+
     let localVarHeaders = this.defaultHeaders;
 
     let localVarHttpHeaderAcceptSelected: string | undefined = options && options.httpHeaderAccept;
@@ -181,12 +231,13 @@ export class ApiPublicOpenapiDataVersionService {
       }
     }
 
-    let localVarPath = `/data-version`;
-    return this.httpClient.request<DataVersion>(
+    let localVarPath = `/comparison-tools/gene-expression`;
+    return this.httpClient.request<Array<GeneExpression>>(
       'get',
       `${this.configuration.basePath}${localVarPath}`,
       {
         context: localVarHttpContext,
+        params: localVarQueryParameters,
         responseType: <any>responseType_,
         withCredentials: this.configuration.withCredentials,
         headers: localVarHeaders,
