+++ libs/model-ad/api-description/src/api.openapi.yaml
@@ -18,14 +18,8 @@ tags:
     description: Operations about the data version.
   - name: Model
     description: Operations about models.
-  - name: Model Overview
-    description: Operations about Model Overview.
   - name: Comparison Tool Config
     description: Operations about Comparison Tool UI configuration.
-  - name: Disease Correlation
-    description: Operations about Disease Correlation results.
-  - name: Gene Expression
-    description: Operations about Gene Expression results.
 paths:
   /data-version:
     $ref: paths/data-version.yaml
@@ -35,12 +29,3 @@ paths:
     $ref: paths/model.yaml
   /comparison-tool-config:
     $ref: paths/comparison-tool-config.yaml
-  /comparison-tools/model-overview:
-    x-sage-internal: true
-    $ref: paths/comparison-tools/model-overview.yaml
-  /comparison-tools/disease-correlation:
-    x-sage-internal: true
-    $ref: paths/comparison-tools/disease-correlation.yaml
-  /comparison-tools/gene-expression:
-    x-sage-internal: true
-    $ref: paths/comparison-tools/gene-expression.yaml
